Simplified app directory
