--- app/dashboard/hooks/use-mark-issue-answered.hook.ts
@@ -1,35 +1,0 @@
-"use client";
-
-import { useProjectIssuesStore } from "../stores/project-issues.store";
-import { Issue } from "../dashboard.types";
-import { CategorizedProjectIssues } from "../dashboard.types";
-import { CONSTANTS } from "@/lib/contants";
-
-export function useMarkIssueAsnwered() {
-  const removeIssue = useProjectIssuesStore((state) => state.removeIssue);
-
-  return (issue: Issue) => {
-
-    let from: keyof CategorizedProjectIssues;
-
-    if (
-      issue.linkedProject ===
-      CONSTANTS.WEB_TEAMS.CORE
-    ) {
-      from = "core";
-    } else if (
-      issue.linkedProject ===
-      CONSTANTS.WEB_TEAMS.LEAP
-    ) {
-      from = "leap";
-    } else if (
-      issue.linkedProject === CONSTANTS.WEB_TEAMS.DEV_WORKFLOW
-    ) {
-      from = "dev";
-    } else {
-      from = "others";
-    }
-
-    removeIssue(from, issue.issueNumber);
-  };
-}
