+++ features/dashboard/shared/unanswered-issues.tab.tsx
@@ -1,15 +1,15 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { IssueCard } from "../components/issue-card";
+import { IssueCard } from "./issues/components/issue-card";
 import { RawIssue } from "@/lib/github/github.types";
 import { LoadingIndicator } from "@/components/layout/loading-indicator";
 import { useLoading } from "@/components/providers/loader-context";
-import { TeamTabs } from "../components/team-tabs";
+import { TeamTabs } from "./team-tabs";
 import { CategorizedProjectIssues, Issue } from "../dashboard.types";
 import { getArchivedIssues } from "../../../db/archived-issues.db";
-import { useProjectIssuesStore } from "../stores/project-issues.store";
-import { categorizeIssues } from "../services/categorize-issues.service";
+import { useProjectIssuesStore } from "./issues/store/project-issues.store";
+import { categorizeIssues } from "./issues/services/categorize-issues.service";
 import { fetchGithubIssues } from "../dashboard.action";
 
 export default function UnansweredIssuesTab() {
